requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.9.0